requirements.txt
PyYAML==6.0.2
jsonschema==4.23.0
python-dateutil==2.9.0.post0
requests==2.32.3
boto3>=1.34.0
